+++ Utils/ImGuiUtils.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Collections.Generic;
+using Dalamud.Utility;
+using ImGuiNET;
+using ImGuiScene;
+using Lumina.Data.Files;
+
+namespace HaselTweaks.Utils;
+
+public static class ImGuiUtils
+{
+    private static readonly Dictionary<int, TextureWrap> icons = new();
+
+    public static void DrawIcon(int iconId, int width = -1, int height = -1)
+    {
+        if (!icons.ContainsKey(iconId))
+        {
+            var tex = Service.Data.GetFile<TexFile>($"ui/icon/{iconId / 1000:D3}000/{iconId:D6}_hr1.tex");
+            if (tex == null)
+                return;
+
+            var texWrap = Service.PluginInterface.UiBuilder.LoadImageRaw(tex.GetRgbaImageData(), tex.Header.Width, tex.Header.Height, 4);
+            if (texWrap.ImGuiHandle == IntPtr.Zero)
+                return;
+
+            icons[iconId] = texWrap;
+        }
+
+        ImGui.Image(icons[iconId].ImGuiHandle, new(width == -1 ? icons[iconId].Width : width, height == -1 ? icons[iconId].Height : height));
+    }
+
+    public static void DrawPaddedSeparator()
+    {
+        ImGui.SetCursorPosY(ImGui.GetCursorPosY() + ImGui.GetStyle().FramePadding.Y);
+        ImGui.Separator();
+        ImGui.SetCursorPosY(ImGui.GetCursorPosY() + ImGui.GetStyle().FramePadding.Y - 1);
+    }
+}